requirements.txt
psutil==5.7.2
sortedcontainers==2.2.2
discord.py==1.4.1
pygount==1.2.3
requests==2.24.0
jsonmerge==1.7.0
pytz==2020.1
sympy==1.6.2
praw==7.1.0
sentry_sdk==0.16.3
num2words==0.5.10
python-dotenv==0.14.0
PyYAML==5.3.1